pyproject.toml
[tool.poetry]
name = "forge"
version = "1.0.0"
description = "AI Development Orchestration System"
authors = ["Forge Team"]
readme = "README.md"
packages = [{include = "forge", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.0"
rich = "^13.0.0"
pydantic = "^2.0.0"
pyyaml = "^6.0.0"
sentence-transformers = "^2.2.0"
numpy = "^1.24.0"
anthropic = "^0.18.0"
httpx = "^0.25.0"
gitpython = "^3.1.0"
tenacity = "^8.2.0"
docker = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
mypy = "^1.5.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
forge = "forge.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/forge --cov-report=term-missing"
